Restructured with MVC architecture.

+++ src/gestock/entity/DairyBaseProduct.java
@@ -6,7 +6,7 @@ public class DairyBaseProduct extends BaseProduct {
 
     private int originAnimal;
 
-    public DairyBaseProduct(int code, String name, String description, String traces, String brand, HashMap<String, Double> nutritionFacts, int originAnimal) {
+    public DairyBaseProduct(long code, String name, String description, String traces, String brand, HashMap<String, Double> nutritionFacts, int originAnimal) {
         super(code, name, description, traces, brand, nutritionFacts);
         this.originAnimal = originAnimal;
     }
