+++ src/gestock/entity/MeatBaseProduct.java
@@ -6,7 +6,7 @@ public class MeatBaseProduct extends BaseProduct {
 
 	private boolean halal;
 
-	public MeatBaseProduct(int code, String name, String description, String traces, String brand, HashMap<String, Double> nutritionFacts, boolean halal) {
+	public MeatBaseProduct(long code, String name, String description, String traces, String brand, HashMap<String, Double> nutritionFacts, boolean halal) {
 		super(code, name, description, traces, brand, nutritionFacts);
 		this.halal = halal;
     }
